fifth

+++ receipts/models.py
@@ -23,7 +23,7 @@ class Receipt(models.Model):
     日付 = models.DateField()
     項目コード = models.ForeignKey(項目リスト,on_delete=models.CASCADE, db_index=True)
     価格 = models.PositiveIntegerField(default=0)
-    登録番号 = models.IntegerField(default=0)
+    登録番号 = models.BigIntegerField(default=0)
     備考 = models.CharField(max_length=255, blank=True, null=True)
     PDF番号 = models.ForeignKey(PDF番号, on_delete=models.CASCADE, db_index=True, related_name='receipts_by_pdf_number')
